fix: resolve G104 json encode errors, G115 int overflow
- Add error handling to json.Encode calls in server.go, adapters.go, health.go, inference.go
- Fix int32 overflow in kv_cache_paged.go by using int64 intermediate
- gosec: 87 -> 84 issues

+++ internal/engine/kv_cache_paged.go
@@ -524,7 +524,7 @@ func (c *PagedKVCache) UpdateBatch(layer int, items []struct {
 			c.blockTables[item.SeqID] = table
 		}
 
-		physPositions[i] = table[logicalBlockIdx]*int32(c.blockSize) + int32(blockOffset)
+		physPositions[i] = int32(int64(table[logicalBlockIdx])*int64(c.blockSize) + int64(blockOffset))
 	}
 	c.mu.Unlock()
